Fixed up some things
Left a comment on a particular lifeline that could break the question (by giving the right answer, but actually causes it to return wrong).  Also added author information to the Lifeline script, proper credit where credit is due!  Lastly: disabled the QuestionScreen object.  This is required due to the way the game loads all the data.

Bug: For some reason the Lifeline script isn't registering the Start function correctly.  Need to look further.